+++ base/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - **BREAKING**: Move metrics from `Agent` trait to `AgentContext`
 - **BREAKING**: Move tracing from `Agent` trait to `AgentContext`
+- **BREAKING**: Move span receive/send logs to base agent.
 - **BREAKING**: Rework configuration using [serde](https://docs.rs/serde)
 - **BREAKING**: Update agent models to match latest specs
 - **BREAKING**: Use [error_chain](https://docs.rs/error-chain) for errors